requirements.txt
Flask==2.3.3
ultralytics==8.1.0
opencv-python==4.9.0.80
numpy
Pillow
gunicorn
requests
